Added csfr token to create update and delete

+++ VolunteerProject/admin/users/delete.php
@@ -1,17 +1,21 @@
 <?php
-include "../functions/connections.php";
 session_start();
-if (!isset($_SESSION[ 'user_login' ])) {
-    header( 'Location: C:/xampp/htdocs/Internship/VolunteerProject/admin/login.php' );
-}
-if (isset($_GET[ "id" ])) {
-    $id     = $_GET[ "id" ];
-    $sql    = "DELETE  FROM volunteerinfos WHERE id = '$id'";
-    $result = $conn->query( $sql );
+include "../functions/connections.php";
 
-    header( "Location: ../allVolunteers.php" );
+if (!isset($_SESSION['user_login'])) {
+    header('Location: C:/xampp/htdocs/Internship/VolunteerProject/admin/login.php');
 }
 
-
-
-?>
+// Check if CSRF token is valid
+if (isset($_GET['csrf_token_delete']) && $_GET['csrf_token_delete'] === $_SESSION['csrf_token_delete']) {
+    unset($_SESSION['csrf_token_delete']);
+    if (isset($_GET["id"])) {
+        $id = $_GET["id"];
+        $sql = "DELETE  FROM volunteerinfos WHERE id = '$id'";
+        $result = $conn->query($sql);
+        header("Location: ../allVolunteers.php");
+    }
+} else {
+    // Invalid token, redirect to error page or show an error message
+    header('Location: ../error.php');
+}
